Added more functionality to main menu.
Added a GameOver, Instructions, and Credits form. The instructions and credits form are accessible from the main menu, while the game over form will return you to the main menu. Currently none of these forms have anything other than a button that will return to the user to the main menu.